requirements.txt
asgiref==3.5.0
distlib==0.3.4
Django==3.2.2
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
filelock==3.6.0
gunicorn==20.1.0
importlib-metadata==4.11.3
Markdown==3.3.6
platformdirs==2.5.1
pytz==2022.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.2.4
virtualenv==20.12.1
yapf==0.31.0
zipp==3.8.0
